Make a base class for stepper motor implementations
